test: strengthen WebKit runtime coverage
