+++ donner/editor/wasm/tests/ios-runtime.spec.ts
@@ -1,4 +1,5 @@
-import { expect, test } from "@playwright/test";
+import { expect, type Page, test } from "@playwright/test";
+import { readCanvasColorStats } from "./canvas-color-stats";
 
 declare global {
   interface Window {
@@ -7,13 +8,29 @@ declare global {
   }
 }
 
-const kFatalRuntimePattern =
-  /Aborted|Assertion failed|RuntimeError|Pthread .* sent an error|WebAssembly runtime unavailable/i;
+interface RuntimeOptions {
+  disableWebGpu?: boolean;
+  renderer?: "tiny_skia";
+}
 
-test("iPhone-profile WebKit starts the TinySkia fallback", async ({ page }) => {
+const kFatalRuntimePattern =
+  /Aborted|Assertion failed|RuntimeError|Pthread .* sent an error|WebAssembly runtime unavailable|Unhandled promise rejection/i;
+
+async function startRuntime(page: Page, options: RuntimeOptions = {}) {
+  if (options.disableWebGpu) {
+    await page.addInitScript(() => {
+      Object.defineProperty(Navigator.prototype, "gpu", {
+        configurable: true,
+        value: undefined,
+      });
+    });
+  }
+
   const baseUrl = process.env.DONNER_WASM_BASE_URL || "http://127.0.0.1:8000";
   const url = new URL(baseUrl);
-  url.searchParams.set("renderer", "tiny_skia");
+  if (options.renderer) {
+    url.searchParams.set("renderer", options.renderer);
+  }
   const fatalMessages: string[] = [];
 
   page.on("console", (message) => {
@@ -28,6 +45,7 @@ test("iPhone-profile WebKit starts the TinySkia fallback", async ({ page }) => {
   await page.goto(url.toString(), { waitUntil: "domcontentloaded" });
   await expect(page.locator("canvas#canvas")).toBeVisible();
   await expect(page.locator("#status")).toBeHidden({ timeout: 30000 });
+  await page.waitForTimeout(2000);
 
   const capabilities = await page.evaluate(() => {
     const probe = document.createElement("canvas");
@@ -38,9 +56,11 @@ test("iPhone-profile WebKit starts the TinySkia fallback", async ({ page }) => {
       backend: window.__donnerBackend,
       canStartWasm: window.__donnerCanStartWasm,
       crossOriginIsolated,
+      hasNavigatorGpu: Boolean(navigator.gpu),
       hasSharedArrayBuffer: typeof SharedArrayBuffer !== "undefined",
       hasWebGl2: Boolean(webGl2),
       isSecureContext,
+      search: window.location.search,
       userAgent: navigator.userAgent,
       viewport: {
         width: editorCanvas?.clientWidth || 0,
@@ -55,8 +75,43 @@ test("iPhone-profile WebKit starts the TinySkia fallback", async ({ page }) => {
   expect(capabilities.hasSharedArrayBuffer).toBe(true);
   expect(capabilities.canStartWasm).toBe(true);
   expect(capabilities.hasWebGl2).toBe(true);
-  expect(capabilities.backend).toBe("tiny_skia");
   expect(capabilities.viewport.width).toBeGreaterThan(0);
   expect(capabilities.viewport.height).toBeGreaterThan(0);
   expect(fatalMessages).toEqual([]);
+
+  return { capabilities, fatalMessages };
+}
+
+test("iPhone-profile WebKit automatically falls back without WebGPU", async ({ page }) => {
+  const { capabilities, fatalMessages } = await startRuntime(page, { disableWebGpu: true });
+
+  expect(capabilities.search).not.toContain("renderer=");
+  expect(capabilities.hasNavigatorGpu).toBe(false);
+  expect(capabilities.backend).toBe("tiny_skia");
+  await page.waitForTimeout(2000);
+  expect(fatalMessages).toEqual([]);
+});
+
+test.describe("WebKit TinySkia presentation", () => {
+  // Mobile layout is intentionally separate work. Use the desktop calibration
+  // at DPR 1 to prove the WebKit renderer path presents document pixels, while
+  // the companion test retains the real iPhone viewport and DPR for startup.
+  test.use({ deviceScaleFactor: 1, viewport: { width: 1600, height: 900 } });
+
+  test("presents document pixels", async ({ page }) => {
+    const { capabilities, fatalMessages } = await startRuntime(page, { renderer: "tiny_skia" });
+    const stats = await readCanvasColorStats(page, {
+      x: 580,
+      y: 80,
+      width: 580,
+      height: 680,
+    });
+
+    expect(capabilities.backend).toBe("tiny_skia");
+    expect(stats.nonBlackPixels, JSON.stringify(stats)).toBeGreaterThan(5000);
+    expect(stats.coloredPixels, JSON.stringify(stats)).toBeGreaterThan(1000);
+    expect(stats.maxChannel, JSON.stringify(stats)).toBeGreaterThan(100);
+    await page.waitForTimeout(1000);
+    expect(fatalMessages).toEqual([]);
+  });
 });
